Improve installation of universal objects
After CAP PR 492, the method FWithGivenX might be installed while neither
F nor X are installed. In this case, adding the universal_object to
recnames does not make sense.
However, in the case that F is available (e.g.
name = info.with_given_without_given_name_pair[1]) we can be sure that X is
available since it can be derived as the source or range of F.