Adjust to https://github.com/homalg-project/CAP_project/pull/1133
This makes two logic templates obsolete. It minimally (< 5%) increases the
runtime of benchmark_FinQuivers_MorphismsOfExternalHom.g due to the
overhead of calling CAP_JIT_INCOMPLETE_LOGIC but is much more general then
the replaced logic templates.

It also noticeably increases the compilation time because the generalized
loop fusion runs in the post-processing step, which re-triggers some of the
compilation phases.