Remove unused logic template
This logic template never matches due to the logic template following it.
This is NOT due to the order of the logic templates (CompilerForCAP tries
to match templates in the order in which they are added), but due to the
compilation process: the syntax trees of the three numbers in the list are
equal at the end of the compilation but differ at the beginning. Thus, the
logic template for non-equal numbers kicks in first. A solution would be to
rewrite `number * number * number` to `number ^ 3`, but in this case
`number * number` also appears in the code and can be reused, so adding
that rewrite rule would not improve things.